+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.3.0 - 2016-10-07
+
+### Added
+
+- The memory limit can now be specified using the `Memory` property -
+
+    ```xml
+    <profile pluginId="cd.go.contrib.elastic-agent.docker" id="foo">
+      <property>
+        <key>Memory</key>
+        <!-- You may use prefix B, KB, MB, GB -->
+        <value>1024MB</value>
+      </property>
+    </profile>
+    ```
+
 ## 0.2.0 - 2016-10-06
 
 ### Fixed